backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
python-dotenv==1.1.0
reportlab==4.4.0
bandit==1.7.10
pip-audit==2.9.0
